# api/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API de delivery Cardapiogo.",
        "title": "Cardapiogo API",
        "contact": {
            "name": "Gustavo Ferreira de Jesus",
            "email": "gustavo081900@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/auth/sign-up": {
            "post": {
                "description": "Sign-Up with local credentials and data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign-Up with local credentials and data",
                "parameters": [
                    {
                        "description": "CreateUserParams",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/user.CreateParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
...
        "/v1/user": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "Bearer Token": []
                    }
                ],
                "description": "User can update your profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User can update your profile",
                "parameters": [
                    {
                        "description": "UpdateProfileParams",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/user.UpdateProfileParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
        "/v1/user/roles": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    },
                    {
                        "Bearer Token": []
                    }
                ],
                "description": "Add a new role for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add a new role for user",
                "parameters": [
                    {
                        "description": "AddRolesParams",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/user.AddRolesParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "user.AddRolesParams": {
            "type": "object",
            "required": [
                "roles"
            ],
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "enum": [
                            "consumer",
                            " delivery_man",
                            " owner",
                            " employee"
                        ],
                        "$ref": "#/definitions/user.Role"
                    }
                }
            }
        },
        "user.CreateParams": {
            "type": "object",
            "required": [
                "email",
                "password",
                "profile"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "*********"
                },
                "profile": {
                    "$ref": "#/definitions/user.CreateProfileParams"
                }
            }
        },
        "user.CreateProfileParams": {
            "type": "object",
            "required": [
                "lastName",
                "name",
                "phone"
            ],
            "properties": {
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "phone": {
                    "type": "string",
                    "example": "13997590579"
                }
            }
        },
        "user.Role": {
            "type": "string",
            "enum": [
                "consumer",
                "owner",
                "employee",
                "delivery_man",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleConsumer",
                "RoleOwner",
                "RoleEmployee",
                "RoleDeliveryMan",
                "RoleAdmin"
            ]
        },
        "user.UpdateProfileParams": {
            "type": "object",
            "required": [
                "lastName",
                "name",
                "phone"
            ],
            "properties": {
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "phone": {
                    "type": "string",
                    "example": "13997590579"
                }
            }
        },
        "xerrors.CustomError": {
            "type": "object",
            "properties": {
                "debug": {},
                "details": {},
                "error": {
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}